+++ gwcs/tags/geometry_models.py
@@ -45,10 +45,12 @@ class SphericalCartesianType(GWCSTransformType):
     @classmethod
     def from_tree_transform(cls, node, ctx):
         transform_type = node['transform_type']
+        wrap_phi_at = node['wrap_phi_at']
+        theta_def = node['theta_def']
         if transform_type == 'spherical_to_cartesian':
-            return SphericalToCartesian()
+            return SphericalToCartesian(wrap_phi_at=wrap_phi_at, theta_def=theta_def)
         elif transform_type == 'cartesian_to_spherical':
-            return CartesianToSpherical()
+            return CartesianToSpherical(wrap_phi_at=wrap_phi_at, theta_def=theta_def)
         else:
             raise TypeError(f"Unknown model_type {transform_type}")
 
@@ -60,5 +62,10 @@ class SphericalCartesianType(GWCSTransformType):
             transform_type = 'cartesian_to_spherical'
         else:
             raise TypeError(f"Model of type {model.__class__} is not supported.")
-        node = {'transform_type': transform_type}
+
+        node = {
+            'transform_type': transform_type,
+            'wrap_phi_at': model.wrap_phi_at,
+            'theta_def': model.theta_def
+        }
         return yamlutil.custom_tree_to_tagged_tree(node, ctx)